requirements.txt
asgiref==3.7.2
Django==4.2.5
django-jazzmin==2.6.0
django-taggit==4.0.0
Pillow==10.0.1
python-dotenv==1.0.0
sqlparse==0.4.4
tzdata==2023.3
